+++ services/rds/src/main/java/com/huaweicloud/sdk/rds/v3/model/RestoreExistingInstanceRequestBodySource.java
@@ -111,7 +111,7 @@ public class RestoreExistingInstanceRequestBodySource {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "restore_time")
 
-    private Integer restoreTime;
+    private Long restoreTime;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "database_name")
@@ -174,7 +174,7 @@ public class RestoreExistingInstanceRequestBodySource {
         this.backupId = backupId;
     }
 
-    public RestoreExistingInstanceRequestBodySource withRestoreTime(Integer restoreTime) {
+    public RestoreExistingInstanceRequestBodySource withRestoreTime(Long restoreTime) {
         this.restoreTime = restoreTime;
         return this;
     }
@@ -183,11 +183,11 @@ public class RestoreExistingInstanceRequestBodySource {
      * 恢复数据的时间点，格式为UNIX时间戳，单位是毫秒，时区为UTC。
      * @return restoreTime
      */
-    public Integer getRestoreTime() {
+    public Long getRestoreTime() {
         return restoreTime;
     }
 
-    public void setRestoreTime(Integer restoreTime) {
+    public void setRestoreTime(Long restoreTime) {
         this.restoreTime = restoreTime;
     }
