+++ services/rds/src/main/java/com/huaweicloud/sdk/rds/v3/model/RestorePoint.java
@@ -31,7 +31,7 @@ public class RestorePoint {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "restore_time")
 
-    private Integer restoreTime;
+    private Long restoreTime;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "database_name")
@@ -89,7 +89,7 @@ public class RestorePoint {
         this.backupId = backupId;
     }
 
-    public RestorePoint withRestoreTime(Integer restoreTime) {
+    public RestorePoint withRestoreTime(Long restoreTime) {
         this.restoreTime = restoreTime;
         return this;
     }
@@ -98,11 +98,11 @@ public class RestorePoint {
      * 恢复数据的时间点，格式为UNIX时间戳，单位是毫秒，时区为UTC。
      * @return restoreTime
      */
-    public Integer getRestoreTime() {
+    public Long getRestoreTime() {
         return restoreTime;
     }
 
-    public void setRestoreTime(Integer restoreTime) {
+    public void setRestoreTime(Long restoreTime) {
         this.restoreTime = restoreTime;
     }
