+++ services/rds/src/main/java/com/huaweicloud/sdk/rds/v3/model/RestoreToExistingInstanceRequestBodySource.java
@@ -111,7 +111,7 @@ public class RestoreToExistingInstanceRequestBodySource {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "restore_time")
 
-    private Integer restoreTime;
+    private Long restoreTime;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "database_name")
@@ -169,7 +169,7 @@ public class RestoreToExistingInstanceRequestBodySource {
         this.backupId = backupId;
     }
 
-    public RestoreToExistingInstanceRequestBodySource withRestoreTime(Integer restoreTime) {
+    public RestoreToExistingInstanceRequestBodySource withRestoreTime(Long restoreTime) {
         this.restoreTime = restoreTime;
         return this;
     }
@@ -178,11 +178,11 @@ public class RestoreToExistingInstanceRequestBodySource {
      * 恢复数据的时间点，格式为UNIX时间戳，单位是毫秒，时区为UTC。
      * @return restoreTime
      */
-    public Integer getRestoreTime() {
+    public Long getRestoreTime() {
         return restoreTime;
     }
 
-    public void setRestoreTime(Integer restoreTime) {
+    public void setRestoreTime(Long restoreTime) {
         this.restoreTime = restoreTime;
     }
